+++ src/core/wrapper/npm-check-updates.js
@@ -4,6 +4,10 @@
  * @author Sébastien Règne
  */
 
+// La dépendance "@types/semver-utils" n'est pas inclue dans le projet
+// "npm-check-updates" alors qu'elle est nécessaire. "@types/semver-utils" a
+// donc été ajoutée dans les dépendances de Metalint.
+// https://github.com/raineorshine/npm-check-updates/issues/1461
 import ncu from "npm-check-updates";
 import Levels from "../levels.js";
 import Severities from "../severities.js";